Fixed Application module logic.
