+++ Projects/Windows/Lua/Lua.vcxproj
@@ -153,12 +153,11 @@
   <ItemGroup>
     <ClInclude Include="..\..\..\Code\Engine\Lua\LuaScriptSystem.h" />
     <ClInclude Include="..\..\..\Code\Engine\Lua\Module.h" />
-    <ClInclude Include="..\..\..\Code\Engine\Lua\ApplicationLuaModule.h" />
     <ClInclude Include="framework.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\Code\Engine\Lua\LuaScriptSystem.cpp" />
-    <ClCompile Include="..\..\..\Code\Engine\Lua\ApplicationLuaModule.cpp" />
+    <ClCompile Include="..\..\..\Code\Engine\Lua\LuaModuleApplication.cpp" />
     <ClCompile Include="dllmain.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
