+++ Projects/Windows/Lua/Lua.vcxproj.filters
@@ -30,9 +30,6 @@
     <ClInclude Include="..\..\..\Code\Engine\Lua\LuaScriptSystem.h">
       <Filter>Header Files\Lua</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\Code\Engine\Lua\ApplicationLuaModule.h">
-      <Filter>Header Files\Lua</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="dllmain.cpp">
@@ -41,7 +38,7 @@
     <ClCompile Include="..\..\..\Code\Engine\Lua\LuaScriptSystem.cpp">
       <Filter>Source Files\Lua</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\Code\Engine\Lua\ApplicationLuaModule.cpp">
+    <ClCompile Include="..\..\..\Code\Engine\Lua\LuaModuleApplication.cpp">
       <Filter>Source Files\Lua</Filter>
     </ClCompile>
   </ItemGroup>
